game_manager_lib

Module scrapers

This module contains functions for scraping game catalogs from various platforms. Each platform has its own submodule that implements the specific scraping logic for that platform.

Módulos:

Amazon Luna

Game Pass PC

Ubisoft+

Re-exports

pub use amazon_luna::fetch_amazon_luna_catalog;

pub use amazon_luna::LunaGame;

pub use ea_play::fetch_ea_play_catalog;

pub use ea_play::EAPlayGame;

pub use game_pass::fetch_game_pass_pc_catalog;

pub use game_pass::GamePassGame;

pub use ubisoft_plus::fetch_ubisoft_plus_catalog;

pub use ubisoft_plus::UbisoftGame;

Modules

amazon_luna

Scraper - Amazon Luna

ea_play

Scraper - EA Play

game_pass

Scraper - Game Pass PC

ubisoft_plus

Scraper - Ubisoft+